modules/handlers.py
"""
Input/output handlers for all supported model types.
"""

import os
from typing import Dict, List, Any, Optional
from pathlib import Path

class BaseHandler:
    """Base class for all model handlers"""
    
    def __init__(self, model_id: str, task: str):
        self.model_id = model_id
        self.task = task
    
    def generate_imports(self) -> str:
        """Generate import statements"""
        return """
import os
import json
import torch
from transformers import AutoTokenizer, AutoModel, pipeline
"""
    
    def generate_setup(self) -> str:
        """Generate model setup code"""
        return f"""
def setup_model():
    model_id = "{self.model_id}"
    tokenizer = AutoTokenizer.from_pretrained(model_id)
    model = AutoModel.from_pretrained(model_id)
    return model, tokenizer
"""
    
    def generate_inference(self) -> str:
        """Generate inference code"""
        raise NotImplementedError
    
    def generate_output_handling(self) -> str:
        """Generate output handling code"""
        return """
def save_output(output, output_path):
    os.makedirs(output_path, exist_ok=True)
    with open(os.path.join(output_path, 'results.json'), 'w') as f:
        json.dump(output, f, indent=2)
"""

class TextHandler(BaseHandler):
    """Handler for text-based models"""
    
    def generate_imports(self) -> str:
        return super().generate_imports() + """
from transformers import AutoTokenizer, AutoModelForSequenceClassification, AutoModelForCausalLM
"""
    
    def generate_inference(self) -> str:
        if self.task == 'text-classification':
            return """
def process_input(text, model, tokenizer):
    inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=True)
    outputs = model(**inputs)
    predictions = torch.softmax(outputs.logits, dim=1)
    return predictions.tolist()
"""
        elif self.task == 'text-generation':
            return """
def process_input(text, model, tokenizer):
    inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=True)
    outputs = model.generate(**inputs, max_length=100, num_return_sequences=1)
    return tokenizer.batch_decode(outputs, skip_special_tokens=True)
"""
        elif self.task == 'translation':
            return """
def process_input(text, model, tokenizer):
    inputs = tokenizer(text, return_tensors="pt", padding=True, truncation=True)
    outputs = model.generate(**inputs)
    return tokenizer.batch_decode(outputs, skip_special_tokens=True)
"""

class ImageHandler(BaseHandler):
    """Handler for image-based models"""
    
    def generate_imports(self) -> str:
        return super().generate_imports() + """
from PIL import Image
import torchvision.transforms as transforms
from transformers import AutoImageProcessor
"""
    
    def generate_inference(self) -> str:
        if self.task == 'image-classification':
            return """
def process_input(image_path, model, processor):
    image = Image.open(image_path)
    inputs = processor(images=image, return_tensors="pt")
    outputs = model(**inputs)
    probs = outputs.logits.softmax(-1)
    return probs.tolist()
"""
        elif self.task == 'object-detection':
            return """
def process_input(image_path, model, processor):
    image = Image.open(image_path)
    inputs = processor(images=image, return_tensors="pt")
    outputs = model(**inputs)
    results = processor.post_process_object_detection(outputs, threshold=0.5)
    
    detections = []
    for score, label, box in zip(results[0]['scores'], results[0]['labels'], results[0]['boxes']):
        detections.append({
            'score': score.item(),
            'label': processor.model.config.id2label[label.item()],
            'box': box.tolist()
        })
    return detections
"""
        elif self.task == 'image-segmentation':
            return """
def process_input(image_path, model, processor):
    image = Image.open(image_path)
    inputs = processor(images=image, return_tensors="pt")
    outputs = model(**inputs)
    
    segmentation = processor.post_process_semantic_segmentation(outputs)
    return segmentation[0].numpy().tolist()
"""

class AudioHandler(BaseHandler):
    """Handler for audio-based models"""
    
    def generate_imports(self) -> str:
        return super().generate_imports() + """
import librosa
import soundfile as sf
from transformers import AutoProcessor
"""
    
    def generate_inference(self) -> str:
        if self.task == 'automatic-speech-recognition':
            return """
def process_input(audio_path, model, processor):
    audio, rate = librosa.load(audio_path)
    inputs = processor(audio, sampling_rate=rate, return_tensors="pt")
    outputs = model(**inputs)
    return processor.batch_decode(outputs.logits.argmax(-1))
"""
        elif self.task == 'text-to-speech':
            return """
def process_input(text, model, processor):
    inputs = processor(text=text, return_tensors="pt")
    speech = model.generate_speech(inputs["input_ids"], processor)
    return speech.numpy().tolist()
"""

class VideoHandler(BaseHandler):
    """Handler for video-based models"""
    
    def generate_imports(self) -> str:
        return super().generate_imports() + """
import decord
import torch.nn.functional as F
from transformers import AutoProcessor
"""
    
    def generate_inference(self) -> str:
        if self.task == 'video-classification':
            return """
def process_input(video_path, model, processor):
    video = decord.VideoReader(video_path)
    frames = video.get_batch(list(range(0, len(video), len(video)//16))).asnumpy()
    inputs = processor(videos=frames, return_tensors="pt")
    outputs = model(**inputs)
    return F.softmax(outputs.logits, dim=1).tolist()
"""
        elif self.task == 'video-generation':
            return """
def process_input(text, model, processor):
    inputs = processor(text, return_tensors="pt")
    video = model.generate(inputs["input_ids"])
    return video.numpy().tolist()
"""

class MultiModalHandler(BaseHandler):
    """Handler for multi-modal models"""
    
    def generate_imports(self) -> str:
        return super().generate_imports() + """
from PIL import Image
from transformers import AutoProcessor
"""
    
    def generate_inference(self) -> str:
        if self.task == 'visual-question-answering':
            return """
def process_input(image_path, question, model, processor):
    image = Image.open(image_path)
    inputs = processor(images=image, text=question, return_tensors="pt")
    outputs = model(**inputs)
    return processor.decode(outputs.logits.argmax(-1))
"""
        elif self.task == 'document-question-answering':
            return """
def process_input(image_path, question, model, processor):
    image = Image.open(image_path)
    inputs = processor(image, question, return_tensors="pt")
    outputs = model(**inputs)
    answer_start = outputs.start_logits.argmax()
    answer_end = outputs.end_logits.argmax()
    return {
        'answer': processor.decode(inputs['input_ids'][0][answer_start:answer_end+1]),
        'start': answer_start.item(),
        'end': answer_end.item()
    }
"""

def get_handler(input_type: str, model_id: str, task: str) -> BaseHandler:
    """Get appropriate handler for model type"""
    handlers = {
        'text': TextHandler,
        'image': ImageHandler,
        'audio': AudioHandler,
        'video': VideoHandler,
        'multimodal': MultiModalHandler
    }
    handler_class = handlers.get(input_type, TextHandler)
    return handler_class(model_id, task)

def get_inference_code(input_type: str, model_id: str, task: str) -> str:
    """Generate complete inference code for model"""
    handler = get_handler(input_type, model_id, task)
    
    code_parts = [
        handler.generate_imports(),
        handler.generate_setup(),
        handler.generate_inference(),
        handler.generate_output_handling(),
        """
if __name__ == '__main__':
    # Setup model and processor
    model, processor = setup_model()
    
    # Get input path from environment
    input_path = os.environ.get('INPUT_PATH', '/workspace/input')
    output_path = os.environ.get('OUTPUT_PATH', '/workspace/output')
    
    # Process all files in input directory
    results = []
    for filename in os.listdir(input_path):
        file_path = os.path.join(input_path, filename)
        try:
            if input_type == 'multimodal':
                # Handle multimodal inputs
                image_path = os.path.join(input_path, filename)
                question = os.environ.get('QUESTION', 'What is in this image?')
                result = process_input(image_path, question, model, processor)
            else:
                # Handle single modality inputs
                input_path = os.path.join(input_path, filename)
                result = process_input(input_path, model, processor)
            
            results.append({
                'filename': filename,
                'result': result
            })
        except Exception as e:
            results.append({
                'filename': filename,
                'error': str(e)
            })
    
    # Save results
    save_output(results, output_path)
"""
    ]
    
    return '\n'.join(code_parts)