requirements.txt
flask>=2.3.3
huggingface_hub>=0.19.4
requests>=2.31.0
transformers>=4.36.0
torch>=2.1.0
safetensors>=0.4.0
accelerate>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
packaging>=23.1
tqdm>=4.66.1
aiohttp>=3.9.1
urllib3>=2.1.0
jsonschema>=4.20.0
validators>=0.22.0